Add a tuner firmware. Test new parameters for LKP-Serial.
The default parameters given by the auto mode is too conservative
for low-latency applications like us. The debouncing and filtering
can increase the time required to decide the electrode status. The
values are now a lot more aggressive on optimizing the decision
latency and the total scan time.

The change above also sacrifices noise immunity since the sensors
are now a lot more sensitive to external environment changes. It is
now advised to have the sensor board AND the overlay taped together
and mounted on a flat surface in order to maximize the chance of
getting false triggering. It may also be a good idea to keep the
slider away from "heavy" EMI sources (e.g. running air conditioners)
so the potential interference can be kept minimum.